+++ packages/mds-jurisdiction-service/service/handlers/delete-jurisdiction-handler.ts
