+++ packages/mds-jurisdiction/tests/api.spec.ts
@@ -18,7 +18,7 @@ import supertest from 'supertest'
 import test from 'unit.js'
 import { ApiServer } from '@mds-core/mds-api-server'
 import { v4 as uuid } from 'uuid'
-import { JurisdictionServiceProvider } from '@mds-core/mds-jurisdiction-service/server'
+import { JurisdictionServiceProvider } from '@mds-core/mds-jurisdiction-service/service/provider'
 import { SCOPED_AUTH } from '@mds-core/mds-test-data'
 import { JurisdictionDomainModel } from '@mds-core/mds-jurisdiction-service'
 import { api } from '../api'
